refactor: move db utilities
