+++ src/pageql/pageql.py
@@ -36,120 +36,17 @@ from pageql.reactive import (
     _convert_dot_sql,
 )
 from pageql.reactive_sql import parse_reactive, _replace_placeholders
+from pageql.database import (
+    connect_database,
+    flatten_params,
+    parse_param_attrs,
+    db_execute_dot,
+    evalone,
+)
 import sqlglot
 
 
-def connect_database(db_path: str):
-    """Return ``(connection, dialect)`` for the given path or URL."""
-    if db_path.startswith("postgres://") or db_path.startswith("postgresql://"):
-        try:
-            import psycopg
-            return psycopg.connect(db_path), "postgresql"
-        except ImportError:
-            try:
-                import psycopg2
-                return psycopg2.connect(db_path), "postgresql"
-            except ImportError:  # pragma: no cover - optional deps
-                raise ImportError(
-                    "PostgreSQL support requires psycopg or psycopg2 to be installed"
-                )
-    if db_path.startswith("mysql://"):
-        parsed = urlparse(db_path)
-        try:
-            import mysql.connector
-            return (
-                mysql.connector.connect(
-                    user=parsed.username,
-                    password=parsed.password,
-                    host=parsed.hostname,
-                    port=parsed.port or 3306,
-                    database=parsed.path.lstrip("/"),
-                ),
-                "mysql",
-            )
-        except ImportError:
-            try:
-                import pymysql
-                return (
-                    pymysql.connect(
-                        host=parsed.hostname,
-                        user=parsed.username,
-                        password=parsed.password,
-                        database=parsed.path.lstrip("/"),
-                        port=parsed.port or 3306,
-                    ),
-                    "mysql",
-                )
-            except ImportError:  # pragma: no cover - optional deps
-                raise ImportError(
-                    "MySQL support requires mysql-connector-python or PyMySQL"
-                )
-    if db_path.startswith("sqlite://"):
-        db_path = db_path.split("://", 1)[1]
-    return sqlite3.connect(db_path), "sqlite"
 
-def flatten_params(params):
-    """
-    Recursively flattens a nested dictionary using __ separator.
-    
-    Args:
-        params: A dictionary, potentially containing nested dictionaries
-        
-    Returns:
-        A flattened dictionary
-        
-    Example:
-        >>> flatten_params({"a": {"b": "c"}})
-        {'a__b': 'c'}
-        >>> flatten_params({"x": 1, "y": {"z": 2, "w": {"v": 3}}})
-        {'x': 1, 'y__z': 2, 'y__w__v': 3}
-    """
-    result = {}
-    for key, value in params.items():
-        if isinstance(value, dict):
-            flattened = flatten_params(value)
-            for k, v in flattened.items():
-                result[f"{key}__{k}"] = v
-        else:
-            result[key] = value
-    return result
-
-def parse_param_attrs(s):
-    """Parse attributes from a ``#param`` directive.
-
-    Handles optional whitespace around the ``=`` sign and quoted values.
-    Tokens with an ``=`` but no following value are treated as boolean flags
-    equivalent to specifying just the name.
-
-    Example::
-
-        >>> parse_param_attrs("a=1 b = 2 c")
-        {'a': '1', 'b': '2', 'c': True}
-        >>> parse_param_attrs("a='hello world' optional")
-        {'a': 'hello world', 'optional': True}
-        >>> parse_param_attrs("a=1 b=2 flag")
-        {'a': '1', 'b': '2', 'flag': True}
-    """
-
-    if not s:
-        return {}
-
-    attrs: dict[str, object] = {}
-    token_re = re.compile(r'(\w+)(?:\s*=\s*("[^"]*"|\'[^\']*\'|\S+))?')
-
-    for match in token_re.finditer(s):
-        key = match.group(1)
-        value = match.group(2)
-        if value is None:
-            attrs[key] = True
-            continue
-        if (value.startswith('"') and value.endswith('"')) or (value.startswith("'") and value.endswith("'")):
-            value = value[1:-1]
-        attrs[key] = value
-
-    return attrs
-
-_DV_CACHE: dict[tuple[int, str, tuple], DerivedSignal2] = {}
 _ONEVENT_CACHE: dict[tuple[int, int, str, tuple], str] = {}
 
 # Short descriptions for valid PageQL directives. Each entry includes a
@@ -264,137 +161,6 @@ class RenderContext:
                 self.scripts.append(content)
 
 
-def db_execute_dot(db, exp, params):
-    """
-    Executes an SQL expression after converting dot notation parameters to double underscore format.
-    
-    Args:
-        db: SQLite database connection
-        exp: SQL expression string
-        params: Parameters dictionary
-        
-    Returns:
-        The result of db.execute()
-        
-    Example:
-        >>> db = sqlite3.connect(":memory:")
-        >>> cursor = db_execute_dot(db, "select :user.name", {"user__name": "John"})
-        >>> cursor.fetchone()[0]
-        'John'
-        >>> cursor = db_execute_dot(db, "select :headers.meta.title", {"headers__meta__title": "Page"})
-        >>> cursor.fetchone()[0]
-        'Page'
-    """
-    # Convert :param.name.subname to :param__name__subname in the expression
-    converted_exp = re.sub(r':([a-zA-Z0-9_]+(?:\.[a-zA-Z0-9_]+)+)',
-                          lambda m: ':' + m.group(1).replace('.', '__'),
-                          exp)
-    param_names = [m.replace('.', '__') for m in re.findall(r':([a-zA-Z0-9_]+(?:\.[a-zA-Z0-9_]+)*)', exp)]
-    missing = [n for n in param_names if n not in params]
-    if missing:
-        avail = ', '.join(sorted(params.keys()))
-        raise ValueError(
-            f"Missing parameter(s) {', '.join(m.replace('__', '.') for m in missing)} "
-            f"for SQL expression `{exp}`. Available parameters: {avail}"
-        )
-
-    converted_params = {}
-    for k, v in params.items():
-        if isinstance(v, Signal):
-            converted_params[k] = v.value
-        else:
-            converted_params[k] = v
-    try:
-        return db.execute(converted_exp, converted_params)
-    except sqlite3.Error as e:
-        raise ValueError(
-            f"Error executing SQL `{converted_exp}` with params {converted_params}: {e}"
-        )
-
-def evalone(db, exp, params, reactive=False, tables=None, expr=None):
-    exp = exp.strip()
-    if exp.upper() == "NULL":
-        if reactive:
-            return ReadOnly(None)
-        return None
-    dialect = getattr(tables, "dialect", "sqlite") if tables is not None else "sqlite"
-    if re.match("^:?[a-zA-z._][a-zA-z._0-9]*$", exp):
-        original = exp[1:] if exp.startswith(":") else exp
-        name = original.replace('.', '__')
-        if name in params:
-            val = params[name]
-            if reactive:
-                if isinstance(val, ReadOnly):
-                    return val
-                if isinstance(val, Signal):
-                    return val
-                signal = DerivedSignal(lambda v=val: v, [])
-                params[name] = signal
-                return signal
-            if isinstance(val, Signal):
-                return val.value
-            return val
-        raise ValueError(
-            f"Missing parameter '{original}' for expression `{exp}`. "
-            f"Available parameters: {', '.join(sorted(params.keys()))}"
-        )
-        
-    if not re.match(r"(?i)^\s*(select|\(select)", exp):
-        exp = "select " + exp
-
-    if reactive:
-        sql = _convert_dot_sql(exp)
-        if tables is None:
-            tables = Tables(db, dialect)
-        dep_names = [name.replace('.', '__') for name in get_dependencies(sql)]
-
-        missing = [n for n in dep_names if n not in params]
-        if missing:
-            avail = ', '.join(sorted(params.keys()))
-            raise ValueError(
-                f"Missing parameter(s) {', '.join(m.replace('__', '.') for m in missing)} "
-                f"for SQL expression `{exp}`. Available parameters: {avail}"
-            )
-
-        for name in dep_names:
-            val = params.get(name)
-            if val is not None and not isinstance(val, Signal):
-                params[name] = ReadOnly(val)
-        deps = []
-        dep_keys = []
-        for name in dep_names:
-            val = params.get(name)
-            if isinstance(val, Signal) and not isinstance(val, ReadOnly):
-                deps.append(val)
-            if isinstance(val, ReadOnly):
-                dep_keys.append(val.value)
-            else:
-                dep_keys.append(id(val))
-        def _build():
-            nonlocal expr
-            if expr is None:
-                expr = sqlglot.parse_one(sql, read=dialect)
-            #print("parse_reactive: ", expr.sql())
-            comp = parse_reactive(expr, tables, params, one_value=True)
-            return comp
-        cache_key = (id(tables), sql, tuple(dep_keys))
-        cache_allowed = "randomblob" not in sql.lower()
-        dv = _DV_CACHE.get(cache_key) if cache_allowed else None
-        if dv is not None:
-            if not hasattr(dv, "listeners") or dv.listeners:
-                return dv
-        dv = derive_signal2(_build, deps)
-        if cache_allowed:
-            _DV_CACHE[cache_key] = dv
-        return dv
-
-    try:
-        r = db_execute_dot(db, exp, params).fetchone()
-        if len(r) != 1:
-            raise ValueError(f"SQL expression `{exp}` with params `{params}` returned {len(r)} rows, expected 1")
-        return r[0]
-    except sqlite3.Error as e:
-        raise ValueError(f"Error evaluating SQL expression `{exp}` with params `{params}`: {e}")
 
 
 class RenderResultException(Exception):
