+++ tests/test_connect_database.py
@@ -2,7 +2,8 @@ import sys
 from pathlib import Path
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent / "src"))
 
-from pageql.pageql import connect_database, PageQL
+from pageql.database import connect_database
+from pageql.pageql import PageQL
 
 
 def test_connect_database_returns_dialect():
