+++ tests/test_evalone.py
@@ -10,7 +10,7 @@ sys.modules["watchfiles"].awatch = lambda *args, **kwargs: None
 
 import pytest
 
-from pageql.pageql import evalone
+from pageql.database import evalone
 from pageql.reactive import DerivedSignal, Tables, Signal
 import sqlglot
